Fix scoping in tests to support Mojolicious 9.25+
The release of Mojolicious 9.25 made changes to Mojolicious::Renderer
such that it now throws when rendering is attempted multiple times.

The tests in t/01-functional.t did not do this intentionally, but did
trigger this condition because they were all using the same
Mojolicious::Lite app with different plugin condifurations.

This commit fixes the test suite by scoping these test classes, which
re-instates support for recent versions of Mojoliciout.

Fixes #34